requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
nordvpn-switcher-plus==0.3.1
pathlib==1.0.1
psutil==5.9.1
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
random-user-agent==1.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.11
Werkzeug==2.2.0
zipp==3.8.1
